requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.5.3
jinja2==3.1.3
python-dotenv==1.0.0
google-generativeai==0.8.3
